moved get output & input prices to packages & fixed test cases
