+++ Cargo.toml
@@ -2,13 +2,13 @@
 resolver = '2'
 
 members = [
-    "packages/*",
-    "contracts/margined_engine",
-    "contracts/margined_fee_pool",
-    "contracts/margined_insurance_fund",
-    "contracts/margined_pricefeed",
-    "contracts/margined_vamm",
-    "contracts/mocks/*",
+  "packages/*",
+  "contracts/margined_engine",
+  "contracts/margined_fee_pool",
+  "contracts/margined_insurance_fund",
+  "contracts/margined_pricefeed",
+  "contracts/margined_vamm",
+  "contracts/mocks/*",
 ]
 
 [workspace.package]
@@ -34,6 +34,7 @@ cw20 = { version = "1.0.1" }
 cw20-base = { version = "1.0.1" }
 cw-multi-test = { git = "https://github.com/oraichain/cw-multi-test.git" } # fix _contract_address attribute
 cw-utils = "1.0.1"
+subtle = "2.5"
 
 margined_utils = { path = "./packages/margined_utils" }
 margined_perp = { path = "./packages/margined_perp" }
