+++ contracts/margined_vamm/src/handle.rs
@@ -4,7 +4,10 @@ use cosmwasm_std::{
 
 use margined_common::{integer::Integer, validate::validate_ratio};
 use margined_perp::margined_vamm::Direction;
-use margined_utils::contracts::helpers::PricefeedController;
+use margined_utils::{
+    contracts::helpers::PricefeedController,
+    tools::price_swap::{get_input_price_with_reserves, get_output_price_with_reserves},
+};
 
 use crate::{
     contract::{
@@ -178,7 +181,7 @@ pub fn swap_input(
 
     let base_asset_amount = if !quote_asset_amount.is_zero() {
         let base_asset_amount = get_input_price_with_reserves(
-            deps.as_ref(),
+            config.decimals,
             &direction,
             quote_asset_amount,
             state.quote_asset_reserve,
@@ -248,7 +251,7 @@ pub fn swap_output(
 
     let quote_asset_amount = if !base_asset_amount.is_zero() {
         let quote_asset_amount = get_output_price_with_reserves(
-            deps.as_ref(),
+            config.decimals,
             &direction,
             base_asset_amount,
             state.quote_asset_reserve,
@@ -357,88 +360,6 @@ pub fn settle_funding(deps: DepsMut, env: Env, info: MessageInfo) -> StdResult<R
         ("underlying_price", &underlying_price.to_string()),
         ("index_price", &index_price.to_string()),
     ]))
-}
-
-pub fn get_input_price_with_reserves(
-    deps: Deps,
-    direction: &Direction,
-    quote_asset_amount: Uint128,
-    quote_asset_reserve: Uint128,
-    base_asset_reserve: Uint128,
-) -> StdResult<Uint128> {
-    if quote_asset_amount == Uint128::zero() {
-        return Ok(Uint128::zero());
-    }
-
-    let config = read_config(deps.storage)?;
-
-    // k = x * y (divided by decimal places)
-    let invariant_k = quote_asset_reserve
-        .checked_mul(base_asset_reserve)?
-        .checked_div(config.decimals)?;
-
-    let quote_asset_after = match direction {
-        Direction::AddToAmm => quote_asset_reserve.checked_add(quote_asset_amount)?,
-        Direction::RemoveFromAmm => quote_asset_reserve.checked_sub(quote_asset_amount)?,
-    };
-
-    let invariant_k_decimals = invariant_k.checked_mul(config.decimals)?;
-
-    let base_asset_after = invariant_k_decimals.checked_div(quote_asset_after)?;
-
-    let mut base_asset_bought = base_asset_after.abs_diff(base_asset_reserve);
-
-    // follows the design of the perpetual protocol decimals
-    // https://github.com/perpetual-protocol/perpetual-protocol/blob/release/v2.1.x/src/utils/Decimal.sol
-    let remainder = invariant_k_decimals.checked_rem(quote_asset_after)?;
-    if remainder != Uint128::zero() {
-        if *direction == Direction::AddToAmm {
-            base_asset_bought = base_asset_bought.checked_sub(1u128.into())?;
-        } else {
-            base_asset_bought = base_asset_bought.checked_add(1u128.into())?;
-        }
-    }
-
-    Ok(base_asset_bought)
-}
-
-pub fn get_output_price_with_reserves(
-    deps: Deps,
-    direction: &Direction,
-    base_asset_amount: Uint128,
-    quote_asset_reserve: Uint128,
-    base_asset_reserve: Uint128,
-) -> StdResult<Uint128> {
-    if base_asset_amount == Uint128::zero() {
-        return Ok(Uint128::zero());
-    }
-    let config = read_config(deps.storage)?;
-
-    let invariant_k = quote_asset_reserve
-        .checked_mul(base_asset_reserve)?
-        .checked_div(config.decimals)?;
-
-    let base_asset_after = match direction {
-        Direction::AddToAmm => base_asset_reserve.checked_add(base_asset_amount)?,
-        Direction::RemoveFromAmm => base_asset_reserve.checked_sub(base_asset_amount)?,
-    };
-
-    let invariant_k_decimals = invariant_k.checked_mul(config.decimals)?;
-
-    let quote_asset_after = invariant_k_decimals.checked_div(base_asset_after)?;
-
-    let mut quote_asset_sold = quote_asset_after.abs_diff(quote_asset_reserve);
-    // follows the design of the perpetual protocol decimals
-    // https://github.com/perpetual-protocol/perpetual-protocol/blob/release/v2.1.x/src/utils/Decimal.sol
-    let remainder = invariant_k_decimals.checked_rem(base_asset_after)?;
-    if remainder != Uint128::zero() {
-        if *direction == Direction::AddToAmm {
-            quote_asset_sold = quote_asset_sold.checked_sub(1u128.into())?;
-        } else {
-            quote_asset_sold = quote_asset_sold.checked_add(1u128.into())?;
-        }
-    }
-    Ok(quote_asset_sold)
 }
 
 pub fn update_reserve(
